+++ src/commands/comments.ts
@@ -6,7 +6,7 @@ import {
   MarkdownString
 } from "vscode";
 import { GistCodeComment } from "../comments";
-import { EXTENSION_ID } from "../constants";
+import { EXTENSION_NAME } from "../constants";
 import {
   createGistComment,
   deleteGistComment,
@@ -33,22 +33,22 @@ async function addComment(reply: CommentReply) {
 
 export function registerCommentCommands(context: ExtensionContext) {
   context.subscriptions.push(
-    commands.registerCommand(`${EXTENSION_ID}.addGistComment`, addComment)
+    commands.registerCommand(`${EXTENSION_NAME}.addGistComment`, addComment)
   );
 
   context.subscriptions.push(
-    commands.registerCommand(`${EXTENSION_ID}.replyGistComment`, addComment)
+    commands.registerCommand(`${EXTENSION_NAME}.replyGistComment`, addComment)
   );
 
   context.subscriptions.push(
     commands.registerCommand(
-      `${EXTENSION_ID}.editGistComment`,
+      `${EXTENSION_NAME}.editGistComment`,
       async (comment: GistCodeComment) => {
         if (!comment.parent) {
           return;
         }
 
-        comment.parent.comments = comment.parent.comments.map(cmt => {
+        comment.parent.comments = comment.parent.comments.map((cmt) => {
           if ((cmt as GistCodeComment).id === comment.id) {
             cmt.mode = CommentMode.Editing;
           }
@@ -60,7 +60,7 @@ export function registerCommentCommands(context: ExtensionContext) {
   );
 
   commands.registerCommand(
-    `${EXTENSION_ID}.saveGistComment`,
+    `${EXTENSION_NAME}.saveGistComment`,
     async (comment: GistCodeComment) => {
       if (!comment.parent) {
         return;
@@ -73,7 +73,7 @@ export function registerCommentCommands(context: ExtensionContext) {
 
       await editGistComment(comment.gistId, comment.id, content);
 
-      comment.parent.comments = comment.parent.comments.map(cmt => {
+      comment.parent.comments = comment.parent.comments.map((cmt) => {
         if ((cmt as GistCodeComment).id === comment.id) {
           cmt.mode = CommentMode.Preview;
         }
@@ -85,7 +85,7 @@ export function registerCommentCommands(context: ExtensionContext) {
 
   context.subscriptions.push(
     commands.registerCommand(
-      `${EXTENSION_ID}.deleteGistComment`,
+      `${EXTENSION_NAME}.deleteGistComment`,
       async (comment: GistCodeComment) => {
         let thread = comment.parent;
         if (!thread) {
@@ -94,7 +94,7 @@ export function registerCommentCommands(context: ExtensionContext) {
 
         await deleteGistComment(comment.gistId, comment.id);
         thread.comments = thread.comments.filter(
-          cmt => (cmt as GistCodeComment).id !== comment.id
+          (cmt) => (cmt as GistCodeComment).id !== comment.id
         );
 
         if (thread.comments.length === 0) {